expanded test to be in their own files and changes directory to allow nosetests to discover tests from default behaviour of the command
