+++ .travis.yml
@@ -8,4 +8,4 @@ python:
   - "pypy"
 install:
   - pip install -r requirements.txt
-script: nosetests tests/*.py
+script: nosetests
